Implemented method for finding roots of quadratic equation
Created QuadraticEquationsSolver class which will represent the app and contain  the methods for calculating the equations roots and logging the results to the console.
Implemented the method called findRoots which is responsible for finding roots of the quadratic equation. The logic of this method if based on general approach of solving quadratic equations, using discriminant