Implemented users input validation for interactive mode (closes #3)
Added the validation of the users input in askForValue method. If the user enters an empty string or not a number value the app will generate an error and ask the user to input the value again until it get a valid one. Also added an extra validation rule to the A  coefficient, so the user won't be able to assign it's value to zero